task_management_system/requirements.txt
asgiref==3.7.2
Django==5.1.1
djangorestframework==3.14.0
django-filter==23.1
sqlparse==0.4.4
pytz==2023.3
tzdata==2023.3


#pip install -r requirements.txt for this project
